requirements.txt
python-dotenv==0.10.1
celery==4.2.1
minio==4.0.6
newspaper3k==0.2.8
